AOS-3965 small rerefactor
